+++ common/src/main/java/com/github/smallinger/copperagebackport/block/CopperGolemStatueBlock.java
@@ -170,6 +170,8 @@ public class CopperGolemStatueBlock extends BaseEntityBlock implements SimpleWat
                         level.removeBlock(pos, false);
                         serverLevel.addFreshEntity(golem);
                         level.playSound(null, pos, ModSounds.COPPER_STATUE_BREAK.get(), SoundSource.BLOCKS, 1.0F, 1.0F);
+                        // TODO: Maybe change particle effect - currently using SCRAPE (3005)
+                        level.levelEvent(null, 3005, pos, 0);
                         level.gameEvent(player, GameEvent.BLOCK_DESTROY, pos);
                         stack.hurtAndBreak(1, player, (p) -> p.broadcastBreakEvent(hand));
                         return InteractionResult.SUCCESS;