+++ fabric/src/main/java/com/github/smallinger/copperagebackport/client/CopperAgeBackportFabricClient.java
@@ -96,7 +96,16 @@ public class CopperAgeBackportFabricClient implements ClientModInitializer {
             ModBlocks.WAXED_COPPER_CHAIN.get(),
             ModBlocks.WAXED_EXPOSED_COPPER_CHAIN.get(),
             ModBlocks.WAXED_WEATHERED_COPPER_CHAIN.get(),
-            ModBlocks.WAXED_OXIDIZED_COPPER_CHAIN.get());
+            ModBlocks.WAXED_OXIDIZED_COPPER_CHAIN.get(),
+            // Copper Grates
+            ModBlocks.COPPER_GRATE.get(),
+            ModBlocks.EXPOSED_COPPER_GRATE.get(),
+            ModBlocks.WEATHERED_COPPER_GRATE.get(),
+            ModBlocks.OXIDIZED_COPPER_GRATE.get(),
+            ModBlocks.WAXED_COPPER_GRATE.get(),
+            ModBlocks.WAXED_EXPOSED_COPPER_GRATE.get(),
+            ModBlocks.WAXED_WEATHERED_COPPER_GRATE.get(),
+            ModBlocks.WAXED_OXIDIZED_COPPER_GRATE.get());
     }
 }